+++ src/config/gro.config.default.ts
@@ -1,13 +1,10 @@
 import type {GroConfigCreator, GroConfigPartial} from './config.js';
 import {ENV_LOG_LEVEL, LogLevel} from '../utils/log.js';
 import {
-	hasDeprecatedGroFrontend,
-	hasApiServer,
 	hasNodeLibrary,
 	PRIMARY_NODE_BUILD_CONFIG,
-	API_SERVER_BUILD_CONFIG,
-	toDefaultBrowserBuild,
 	NODE_LIBRARY_BUILD_CONFIG,
+	hasSvelteKitFrontend,
 } from '../build/defaultBuildConfig.js';
 
 /*
@@ -16,6 +13,8 @@ This is the default config that's used
 if the current project does not define one at `src/gro.config.ts`.
 It looks at the project and tries to do the right thing:
 
+- if `src/routes` and `src/app.html`,
+	assumes a SvelteKit frontend
 - if `src/index.html` and `src/index.ts`,
 	assumes Gro's deprecated SPA mode - https://github.com/feltcoop/gro/issues/106
 - if `src/index.ts` and not `src/index.html`,
@@ -26,29 +25,38 @@ It looks at the project and tries to do the right thing:
 */
 
 export const config: GroConfigCreator = async ({fs}) => {
-	const [enableGroFrontend, enableNodeLibrary, enableApiServer] = await Promise.all([
-		hasDeprecatedGroFrontend(fs),
+	const [
+		enableNodeLibrary,
+		// enableApiServer,
+		enableSvelteKitFrontend,
+		// enableGroFrontend,
+	] = await Promise.all([
 		hasNodeLibrary(fs),
-		hasApiServer(fs),
+		// hasApiServer(fs),
+		hasSvelteKitFrontend(fs),
+		// hasDeprecatedGroFrontend(fs),
 	]);
 	const partial: GroConfigPartial = {
 		builds: [
 			PRIMARY_NODE_BUILD_CONFIG,
-			enableGroFrontend ? toDefaultBrowserBuild() : null, // TODO configure asset paths
 			enableNodeLibrary ? NODE_LIBRARY_BUILD_CONFIG : null,
-			enableApiServer ? API_SERVER_BUILD_CONFIG : null,
+			// enableApiServer ? API_SERVER_BUILD_CONFIG : null,
+			// enableGroFrontend ? toDefaultBrowserBuild() : null, // TODO configure asset paths
+			// note there's no build for SvelteKit frontends - should there be?
 		],
 		logLevel: ENV_LOG_LEVEL ?? LogLevel.Trace,
-		// TODO should this use `defaultAdapt`? might be different?
 		adapt: async () => {
 			return [
-				// TODO
-				// enableApiServer ? (await import('../adapt/gro-adapter-api-server.js')).createAdapter() : null,
-				enableGroFrontend
-					? (await import('../adapt/gro-adapter-spa-frontend.js')).createAdapter()
-					: null,
 				enableNodeLibrary
 					? (await import('../adapt/gro-adapter-node-library.js')).createAdapter()
+					: null,
+				// TODO
+				// enableApiServer ? (await import('../adapt/gro-adapter-api-server.js')).createAdapter() : null,
+				// enableGroFrontend
+				// 	? (await import('../adapt/gro-adapter-spa-frontend.js')).createAdapter()
+				// 	: null,
+				enableSvelteKitFrontend
+					? (await import('../adapt/gro-adapter-sveltekit-frontend.js')).createAdapter()
 					: null,
 			];
 		},